net: fix an addUser method/querystring/body bug
Key-value pairs were being POSTed (correct by Spotify's spec) for
action=addUser but they were in the query string rather than in the
x-www-form-urlencoded body. This caused librespot-java 1.6.3 to refuse
connection as it could not find the action type (addUser) because it
wasn't in the body. The official Spotify client uses the body of the
POST. The rust librespot application must be lenient in this regard.

The added urlencoding dependency was already a transitive dependency
of the minreq create with 'urlencoding' feature.